requirements.txt
numpy
pandas
sqlalchemy
psycopg2-binary
ipykernel
requests
tqdm
GeoAlchemy2
geopandas